requirements.txt
asgiref==3.8.1
attrs==24.3.0
babel==2.16.0
certifi==2024.12.14
charset-normalizer==3.4.0
coverage==7.6.9
distlib==0.3.9
Django==5.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docxcompose==1.4.0
docxtpl==0.19.0
drf-spectacular==0.28.0
filelock==3.16.1
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.2
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
python-docx==1.1.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
ruff==0.8.3
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.28.0
